+++ tech1-framework-b2b-mongodb-security-jwt/src/main/java/io/tech1/framework/b2b/mongodb/security/jwt/services/BaseSuperAdminService.java
@@ -5,7 +5,7 @@ import io.tech1.framework.b2b.mongodb.security.jwt.domain.dto.responses.Response
 
 import java.util.List;
 
-public interface SuperAdminService {
+public interface BaseSuperAdminService {
     List<ResponseInvitationCode1> findUnused();
     ResponseServerSessionsTable getServerSessions();
 }