+++ tech1-framework-configurations/src/main/java/io/tech1/framework/configurations/events/ApplicationEvents.java
@@ -8,10 +8,14 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.event.ApplicationEventMulticaster;
 import org.springframework.context.event.SimpleApplicationEventMulticaster;
 import org.springframework.core.task.SimpleAsyncTaskExecutor;
+import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
+import org.springframework.scheduling.support.TaskUtils;
 
 import javax.annotation.PostConstruct;
 
 import static io.tech1.framework.domain.properties.utilities.PropertiesAsserter.assertProperties;
+import static io.tech1.framework.domain.utilities.processors.ProcessorsUtility.getHalfOfCores;
+import static io.tech1.framework.domain.utilities.processors.ProcessorsUtility.getNumOfCores;
 
 @Configuration
 @RequiredArgsConstructor(onConstructor = @__(@Autowired))
@@ -28,10 +32,14 @@ public class ApplicationEvents {
     @Bean(name = "applicationEventMulticaster")
     public ApplicationEventMulticaster simpleApplicationEventMulticaster() {
         var threadNamePrefix = this.applicationFrameworkProperties.getEventsConfigs().getThreadNamePrefix();
-        var simpleAsyncTaskExecutor = new SimpleAsyncTaskExecutor();
-        simpleAsyncTaskExecutor.setThreadNamePrefix(threadNamePrefix);
+        var taskExecutor = new ThreadPoolTaskExecutor();
+        taskExecutor.setThreadNamePrefix(threadNamePrefix);
+        taskExecutor.setCorePoolSize(getHalfOfCores());
+        taskExecutor.setMaxPoolSize(getNumOfCores());
+        taskExecutor.initialize();
         var eventMulticaster = new SimpleApplicationEventMulticaster();
-        eventMulticaster.setTaskExecutor(simpleAsyncTaskExecutor);
+        eventMulticaster.setTaskExecutor(taskExecutor);
+        eventMulticaster.setErrorHandler(TaskUtils.LOG_AND_SUPPRESS_ERROR_HANDLER);
         return eventMulticaster;
     }
 }